change golang code position from off-chain directory to outside
