+++ blockchain/browsersync/goerli/goerliEventLogScan.go
@@ -2,13 +2,13 @@ package goerli
 
 import (
 	"math/big"
-	"payment-bridge/scan/blockchain/goerliclient"
-	"payment-bridge/scan/common/constants"
-	common2 "payment-bridge/scan/common/utils"
-	"payment-bridge/scan/config"
-	"payment-bridge/scan/database"
-	"payment-bridge/scan/logs"
-	models2 "payment-bridge/scan/models"
+	"payment-bridge/blockchain/goerliclient"
+	"payment-bridge/common/constants"
+	common2 "payment-bridge/common/utils"
+	"payment-bridge/config"
+	"payment-bridge/database"
+	"payment-bridge/logs"
+	models2 "payment-bridge/models"
 	"strconv"
 	"sync"
 	"time"
