+++ blockchain/goerliclient/goerliClient.go
@@ -1,11 +1,11 @@
-package polygonclient
+package goerliclient
 
 import (
 	"context"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"math/big"
-	"payment-bridge/scan/logs"
+	"payment-bridge/logs"
 )
 
 /**
