+++ config/config.go
@@ -43,8 +43,7 @@ var config *Configuration
 
 func InitConfig(configFile string) {
 	if strings.Trim(configFile, " ") == "" {
-		configFile = "./scan/config/config.toml"
-		//configFile = "./config/config.toml"
+		configFile = "./config/config.toml"
 	}
 	if metaData, err := toml.DecodeFile(configFile, &config); err != nil {
 		log.Fatal("error:", err)
