+++ database/database.go
@@ -4,8 +4,8 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/mysql"
-	"payment-bridge/scan/config"
-	"payment-bridge/scan/logs"
+	"payment-bridge/config"
+	"payment-bridge/logs"
 )
 
 type Database struct {
