+++ main/payment-bridge.go
@@ -3,15 +3,15 @@ package main
 import (
 	"github.com/gin-gonic/gin"
 	cors "github.com/itsjamie/gin-cors"
-	"payment-bridge/scan/blockchain/browsersync/goerli"
-	"payment-bridge/scan/blockchain/browsersync/polygon"
-	"payment-bridge/scan/blockchain/goerliclient"
-	polygonclient "payment-bridge/scan/blockchain/polygonclient"
-	"payment-bridge/scan/common/constants"
-	"payment-bridge/scan/config"
-	"payment-bridge/scan/database"
-	"payment-bridge/scan/logs"
-	"payment-bridge/scan/routers"
+	"payment-bridge/blockchain/browsersync/goerli"
+	"payment-bridge/blockchain/browsersync/polygon"
+	"payment-bridge/blockchain/goerliclient"
+	polygonclient "payment-bridge/blockchain/polygonclient"
+	"payment-bridge/common/constants"
+	"payment-bridge/config"
+	"payment-bridge/database"
+	"payment-bridge/logs"
+	"payment-bridge/routers"
 	"time"
 )
